pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0b2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "graph_hdc"
version = "0.1.0"
description = "Insert Description"
license = "MIT license"
authors = ["Jonas Teufel <jonseb1998@gmail.com>"]
maintainers = ["Jonas Teufel <jonseb1998@gmail.com>"]
readme = "README.rst"
keywords = []
packages = [
    { include = "graph_hdc" },
    { include = "graph_hdc/experiments"},
]
include = [
    "graph_hdc/VERSION",
]

[tool.poetry.scripts]
graph_hdc = 'graph_hdc.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
pycomex = ">=0.9.2"
click = ">=7.1.2"
jinja2 = ">=3.0.3"
matplotlib = ">=3.5.3"
numpy = ">=1.23.2"
python-decouple = ">=3.6"
poetry-bumpversion = ">=0.3.0"
jsonpickle = ">=4.0.1,<=5.0.0"
torch = ">=2.5.0,<3.0"
torch_geometric = ">=2.6.1,<3.0"
pytorch_lightning = ">=2.4.0,<3.0"
rich_click = ">=1.8.9,<=1.9.0"
seaborn = ">=0.12.2"
umap-learn = ">=0.5.3"
imblearn = ">=0.0"
normflows = ">=1.7.3"
chem_mat_database = { git = "https://github.com/the16thpythonist/chem_mat_data.git" }

[tool.poetry.dev-dependencies]
pytest = ">=7.1.3"
poetry-bumpversion = ">=0.3.0"

[tool.poetry_bumpversion.file."graph_hdc/VERSION"]
search = "{current_version}"
replace = "{new_version}"

[tool.poetry_bumpversion.file."README.rst"]
search = 'version-{current_version}-orange.svg'
replace = 'version-{new_version}-orange.svg'

[tool.poetry_bumpversion.file."graph_hdc/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'